Update 6 -- almost complete Eller Implementation
The different parts of the Eller algorithm are now implemented:
    * intialization of cells with distinct set values, handling rooms
    * randomly removing walls within rows
    * propagation of sets across rows
    * joining of all sets upon reaching last row

The only remaining thing to do is to check the final cells array and
make sure that no rooms are left over which were not accessed; if this
happens, they must be opened up. I should be able to do this pretty
simply with the methods I have in place.

Current progress is sufficient to call it a night -- until tomorrow.