Update 7 -- Eller implemented & Run
Eller's algorithm is functional and embedded in the maze application;
the application works for various difficulty levels. Still some sanity
checking to do. Relevant files:
    *MazeBuilderEller
    *MazeBuilder
    *mazeApplication

Some other files are updated here either because I added notes to
documentation or was fiddling with print statements. Should have
committed/pushed these files immediately (or kept a log of changes to
see if any real functional changes were made--should not be so)--a
lesson for next time.