Update 10 (Release 1.0) -- revised handling of Eller corner cases
Some testing led me to discover that the corner case of leftover rooms
in Eller's algorithm was not being handled correctly in prior
implementation, so I corrected this.

Other things:
   * Updated javadoc/commenting
   * Additional testing for Eller's algorithm
   * checked coverage with EclEmma

Since things are in working order and sufficiently complete, tagging
this as Release 1.0. Any other updates are made between now and tomorrow
will comprise Release 2.0.