P3 -- commit 3.1: data flow between classes
The BasicRobot and Controller classes now exchange data and are mutually
aware. BasicRobot is aware of Controller, its currentState, and the Maze
and associated data (Floorplan, Distance, etc.)

Due to the requirement that the Robot interface is meant to interact
with a game in a playing state, the BasicRobot uses a separate class,
Instantiator (implementor of Runnable), to wait for the Controller to
transition from generating to playing state before accessing relevant
Maze data. Implementation may develop going forward (in particular,
there may still be room for race conditions once the robot is integrated
into the GUI), but commit now for initial working version.