P3 -- commit 3.5: move/rotate and requisites
The robot's move and rotate methods are implemented and integrated into
StatePlaying. They are optimized so as to avoid repeated distance
calculations.

Backwards movements on the part of StatePlaying.keyDown are not yet
handled, and using the down key will lead to incorrect results in the
distance calculations.

Otherwise changes made in the distance calculations are behaving as
expected for the move and rotate methods.

Distance functionality not yet implemented with jump.