P3 -- commit 3.7: redirected control flow
Before it was the case that StatePlaying was directing BasicRobot when
to perform operations. Now the control flow is reversed, and BasicRobot
directs Controller about when StatePlaying should change its internal
state and alter the user interface.

In order to accommodate this, BasicRobot now updates its
position/direction internally without relying on Controller. This offers
the potential for assertion testing: BasicRobot and Controller should
always agree on position & direction.

A boolean field is created in MazeApplication to tell whether the game
should be played with a robot or without. This allows SimpleKeyListener
to give control to the robot when enabled, and directly to Controller
when no robot is invoked.