P3 -- commit 3.10: pseudocode & preliminary code
Basic test cases are pseudocoded in BasicRobotTest; the implementation
will have to have a (simple) way of configuring the application to
generate a fully operational Controller with a robot configured.

I also had a few spare moments to code a maze exploration algorithm for
the robot, which might prove useful in a future project. For now it is
commented out because it is not yet used.