P3 -- committ 3.11: testing development + code refactoring
For the sake of organization and efficiency, general-purpose calculation
function related to the Maze class and its various components have been
delegated to a new class 'MazeMath' in a new package 'comp' (short for
computation).

BasicRobotTest is further developed, with additional helper classes to
aid in testing. These might prove useful for refactoring into code for
RobotDriver in project 4, as they allow for external manipulation of the
Robot interface.