P3 -- commit 3.16: finished testing, increased coverage
More tests added for MazeMath, and new test method added for
ExtendedList functions.

Running coverage check revealed that there were more tests to add for
BasicRobot, which is done.

Coverage for BasicRobot is now ~85% when tested in BasicRobotTest.
Uncovered code mostly comes from exceptions which are not thrown (some
of which should never be thrown, but which are in place to prevent a
compiler error).

Next commit will contain javadoc.