P4 -- commit 4.3: AbstractRobotDriver development
AbstractRobotDriver is still developing; here I have implemented methods
to handle the logic of finding a working sensor that the robot can use
to gauge distances. I am pursuing this first instead of a State pattern
because the code i have written in MazeMath highly simplifies this
logic, which might eliminate the need for a state pattern.

Also included files affected by refactoring in MazeMath.