P4 -- commit 4.4: WallFollower revised
Before, WallFollower and its robot relied on the same sense of relative
direction. This was not compatible with the robot having variable sensor
malfunctions and having to rotate to fill in missing distances at
arbitrary times.

Thus, WallFollower is revised to be agnostic to the robot's directions.
It maintains a sense of absolute direction, and translates this into
relative directions that the robot can follow.

The robot and WallFollower both agree on position.