P4 -- commit 4.9: progress towards sensor buttons
Sensor buttons are provided to StatePlaying; however, they disappear
from view when the MazePanel is updated. Bug to fix.

RobotSensorTrigger implements the mechanism to trigger sensor failure
and repairs. It subclasses Runnable to allow for asynchronicity. Basic
testing in the GUI environment suggests that it is properly functional.
To do: update triggerUpdateSensorInformation in WallFollower to respond
to sensor state.