P4 -- commit 4.10: sensor buttons implemented, interruption handling
I expanded the MazeApplication window to permit space for the sensor
buttons panel, which avoids overlap and thus does not cause the panel to
be drawn over in the playing state. The buttons trigger the sensor
operations as expected.

I also have included mechanisms to allow the robot driver to be
interrupted. If a key is pressed when a driver is present, and the key
does not concern the view of the maze map, it causes immediate
termination of the algorithm and return to the title screen. This is a
simplistic model, but better than having the sorts of errors that may
arise with this issue unchecked.

To permit this change, the RobotDriver interface now specifies an
<interrupt> method.