P5 -- commit 5.13: final tidying & javadoc
Biggest change concerns MazeApplication.java, where classes in the file
were moved into the class as nested classes. For sake of encapsulation,
only MPanel is public; the other three nested classes
(SensorButtonPanel, StartPanel, dButton in MazeApplication are private.
Other files adjusted accordingly.

In MazeColor, I added references to specific methods in java.awt.Color
from which code was reused. The getRGB method is renamed to toInt to
clarify its purpose; a link to Color.getRGB is provided to clarify
source.

Minor comments/adjustments to other files.

All tests pass and application is in working order, so tagging this as
Release 5.0.